Auto detect file type.
Say goodbye to ExcelType!

Awk() throws three kinds of exception:

1) EncryptedDocumentException
When xlsawk detects an *open* encrypted xls(x) file, it throws
an EncryptedDocumentException. But it miss the *edit* encrypted file.
Instead, it throws an IOException because the xls(x) file can't be
decoded normally.

2) InvalidFormatException
Try to open a non-excel file.

3) IOException
Any exception that is not EncryptedDocumentException and
InvalidFormatException.